+++ src/index.js
@@ -50,6 +50,7 @@ export { RowanFormField } from "./form-field/form-field.js";
 export { RowanFormLayout } from "./form-layout/form-layout.js";
 export { RowanFormWizard } from "./form-wizard/form-wizard.js";
 export { RowanIconButton } from "./icon-button/icon-button.js";
+export { RowanImage } from "./image/image.js";
 export { RowanKpiCard } from "./kpi-card/kpi-card.js";
 export { RowanLink } from "./link/link.js";
 export { RowanListbox } from "./listbox/listbox.js";