+++ types/image/index.d.ts
@@ -1,0 +1,1 @@
+export { RowanImage } from "./image.js";